Conflicts:
	mapper/src/main/AndroidManifest.xml

+++ mapper/src/main/java/com/stefano/andrea/tasks/DeleteTask.java
@@ -7,12 +7,12 @@ import android.net.Uri;
 import android.os.AsyncTask;
 
 import com.stefano.andrea.activities.R;
-import com.stefano.andrea.helpers.CommonAlertDialog;
 import com.stefano.andrea.models.Citta;
 import com.stefano.andrea.models.Foto;
 import com.stefano.andrea.models.Posto;
 import com.stefano.andrea.models.Viaggio;
 import com.stefano.andrea.providers.MapperContract;
+import com.stefano.andrea.utils.CommonAlertDialog;
 
 import java.io.File;
 import java.util.List;
@@ -40,7 +40,6 @@ public class DeleteTask<T> extends AsyncTask<Integer, Void, Integer> {
 
     private ContentResolver mResolver;
     private DeleteAdapter mAdapter;
-    private DeleteInterface mDelegate;
     private List<T> mList;
     private List<Integer> mSelectedItems;
     private Activity mActivity;
@@ -57,6 +56,7 @@ public class DeleteTask<T> extends AsyncTask<Integer, Void, Integer> {
     protected Integer doInBackground(Integer... params) {
         if (params.length == 1) {
             //assegno il delegate corretto
+            DeleteInterface mDelegate;
             switch (params[0]) {
                 case CANCELLA_VIAGGIO:
                     mDelegate = new CancellaViaggio();
@@ -128,9 +128,11 @@ public class DeleteTask<T> extends AsyncTask<Integer, Void, Integer> {
     private class CancellaFoto implements DeleteInterface<Foto> {
         @Override
         public int cancellaItem(Foto item) {
-            //cancello il file
-            File foto = new File(item.getPath().substring(7));
-            boolean res = foto.delete();
+            //cancello il file se scattato dall'app
+            if (item.getPath().contains("Mapper")) {
+                File foto = new File(item.getPath().substring(7));
+                boolean res = foto.delete();
+            }
             //cancello il riferimento dal database
             Uri uri = ContentUris.withAppendedId(MapperContract.Foto.CONTENT_URI, item.getId());
             return mResolver.delete(uri, null, null);
